Modernize subscribe redux
