+++ src/redux/actions/index.ts
@@ -7,7 +7,6 @@ import { routingActions } from './routing';
 import { scheduleActions } from './schedule';
 import { sessionsActions } from './sessions';
 import { speakersActions } from './speakers';
-import { subscribeActions } from './subscribe';
 import { teamActions } from './team';
 import { userActions } from './user';
 
@@ -20,7 +19,6 @@ export {
   scheduleActions,
   sessionsActions,
   speakersActions,
-  subscribeActions,
   teamActions,
   userActions,
 };
