+++ src/redux/subscribe/state.ts
@@ -1,0 +1,3 @@
+import { SubscribeState } from './types';
+
+export const initialSubscribeState: SubscribeState = false;
