+++ src/redux/subscribe/types.ts
@@ -1,0 +1,10 @@
+export const SUBSCRIBE = 'SUBSCRIBE';
+
+export type SubscribeState = boolean;
+
+interface SubscribeAction {
+  type: typeof SUBSCRIBE;
+  subscribed: boolean;
+}
+
+export type SubscribeActionTypes = SubscribeAction;
